#66 Nhan Nguyen - divide component
