+++ src/pages/Dashboard/components/DashboardCardList/DashboardCardElement/index.ts
@@ -1,0 +1,3 @@
+import DashboardCardElement from './DashboardCardElement'
+
+export default DashboardCardElement
