+++ src/pages/Dashboard/components/DashboardCardList/index.ts
@@ -1,0 +1,3 @@
+import DashboardCardList from './DashboardCardList'
+
+export default DashboardCardList
