+++ src/pages/Dashboard/components/DashboardTitle/index.ts
@@ -1,0 +1,3 @@
+import DashboardTitle from './DashboardTitle'
+
+export default DashboardTitle
